pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finisher"
dynamic = ["version"]
description = "AI Image Upscaling Tool using Automatic1111 API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "illuminatianon", email = "me@illuminati.com"}
]
keywords = ["ai", "image", "upscaling", "automatic1111", "stable-diffusion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "pyside6>=6.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
finisher = "finisher.main:main"

[project.urls]
Homepage = "https://github.com/illuminatianon/finisher"
Repository = "https://github.com/illuminatianon/finisher"
Issues = "https://github.com/illuminatianon/finisher/issues"

[tool.hatch.version]
path = "src/finisher/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/finisher"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | env
  | ENV
  | \.env
  | site-packages
  | node_modules
  | build
  | dist
  | \.pytest_cache
  | htmlcov
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "ENV",
    ".env",
    "site-packages",
    "node_modules",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    "htmlcov",
    ".mypy_cache",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "venv/",
    ".venv/",
    "env/",
    "ENV/",
    ".env/",
    "site-packages/",
    "node_modules/",
    "build/",
    "dist/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
